Implement undo delete logic

+++ src/context/TaskContext.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo, useContext } from 'react';
+import React, { useState, useEffect, useMemo, useContext, useRef } from 'react';
 import useToastify from '../hooks/useToastify';
 
 import { trackPromise } from 'react-promise-tracker';
@@ -10,7 +10,7 @@ TaskContext.displayName = 'TaskContext';
 const TaskProvider = ({ children }) => {
     const [tasks, setTasks] = useState([]);
     const [deletedTasks, setDeletedTasks] = useState([]);
-    const [timeoutIds, setTimeoutIds] = useState([]);
+    const isDeleting = useRef(false);
 
     const { notifySuccess, notifyDelete } = useToastify();
 
@@ -21,21 +21,15 @@ const TaskProvider = ({ children }) => {
             }),
             'main-area'
         );
-    }, []);
 
-    useEffect(() => {
         return () => {
-            timeoutIds.forEach((t) => {
-                clearTimeout(t);
-            });
-            setTimeoutIds([]);
-
+            console.log('here');
             deletedTasks.forEach((d) => {
                 taskService.deleteTask(d.id);
             });
             setDeletedTasks([]);
         };
-    }, [tasks]);
+    }, []);
 
     const addTask = (taskObj) => {
         taskService.createTask(taskObj).then((id) => {
@@ -44,42 +38,49 @@ const TaskProvider = ({ children }) => {
         });
     };
 
-    const undo = () => {
-        const newDeletedTasks = [...deletedTasks];
-        const latestDeletedTask = newDeletedTasks.pop();
-        setDeletedTasks(newDeletedTasks);
+    const deleteTask = (task) => {
+        isDeleting.current = true;
+        const { id } = task;
 
-        const newTimeotIds = [...timeoutIds];
-        newTimeotIds.pop();
-        setTimeoutIds(newTimeotIds);
+        const filteredTasks = tasks.filter((t) => t?.id !== id);
+        setTasks(filteredTasks);
 
-        const newTasks = [...tasks];
-        newTasks.splice(latestDeletedTask.index, 0, latestDeletedTask);
-        setTasks(newTasks);
-    };
+        const isAlreadyDeleted = deletedTasks.some((t) => t?.id === id);
 
-    const deleteTask = (id) => {
-        const newTasks = [...tasks];
-
-        let toDelete = newTasks.find((t) => t.id === id);
-        const toDeleteIndex = newTasks.indexOf(toDelete);
-
-        newTasks.splice(toDeleteIndex, 1);
-        setTasks(newTasks);
-
-        if (!deletedTasks.includes(toDelete)) {
-            setDeletedTasks((prev) => [
-                ...prev,
-                { ...toDelete, index: toDeleteIndex },
-            ]);
+        let newDeletedTasks = [];
+        if (!isAlreadyDeleted) {
+            newDeletedTasks = [...deletedTasks, { ...task }];
+            setDeletedTasks(newDeletedTasks);
         }
 
-        const currTimeoutId = setTimeout(() => {
-            taskService.deleteTask(id);
-        }, 10000);
-        setTimeoutIds((prev) => [...prev, currTimeoutId]);
+        notifyDelete(
+            'Task Deleted',
+            () => undo(task, newDeletedTasks),
+            () => deletePermanently(id)
+        );
+    };
 
-        notifyDelete('Task Deleted', undo);
+    const deletePermanently = (id) => {
+        console.log(isDeleting.current);
+        if (isDeleting.current) {
+            taskService.deleteTask(id);
+        }
+    };
+
+    const undo = (task, deletedTasks) => {
+        isDeleting.current = false;
+
+        // We can mutate the input array here because
+        // it's a copy of the state array
+        const latestDeletedTask = deletedTasks.pop();
+        setDeletedTasks(deletedTasks);
+
+        const newTasks = [...tasks];
+        // The task is also deleted here although it had
+        // already been deleted in deleteTask() because
+        // the state wouldn't have been updated yet on undo call
+        newTasks.splice(task?.index, 1, task);
+        setTasks(newTasks);
     };
 
     const value = useMemo(
